// src/generated/api-types.ts
export interface components {
    schemas: {
// ...
        ProductResponseDto: {
            id: string;
            name: string;
            image?: string;
            description?: string;
            price: number;
            available: boolean;
        };
        CreateProductDto: {
            name: string;
            price: number;
            image?: string;
            description?: string;
            available?: boolean;
        };
        UpdateProductDto: {
            name?: string;
            price?: number;
            image?: string;
            description?: string;
            available?: boolean;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}